shared module for cli args
